requirements.txt
Flask
requests
sqlite3
line-bot-sdk